Reapply "Mejora en controlador de calculadora y validaciones de tipos"
This reverts commit 2ef2edb15f81e36db9121b1cc7f6725e2d4505ce.

+++ backend/src/main/java/com/dcplatform/api/calculator/CalculatorController.java
@@ -15,13 +15,11 @@ public class CalculatorController {
 
     @PostMapping("/estimate")
     public ResponseEntity<CalculatorResponse> estimate(@Valid @RequestBody CalculatorRequest request) {
-        BigDecimal carga = request.getCargaTermicaKw();
-        
-        // Lógica de estimación básica
+        BigDecimal carga = BigDecimal.valueOf(request.getCargaTermicaKw());
+
         BigDecimal consumoEstimadoKwh = carga.multiply(new BigDecimal("720")); 
         BigDecimal costoEstimado = consumoEstimadoKwh.multiply(new BigDecimal("0.15")); 
 
-        // Creamos la respuesta usando el constructor con los dos argumentos
         CalculatorResponse response = new CalculatorResponse(consumoEstimadoKwh, costoEstimado);
 
         return ResponseEntity.ok(response);
